+++ include/Metrics.h
@@ -17,13 +17,13 @@ private:
   public:
     int id;
     std::string tag;
-    int currentValue;
-    int minValue;
-    int maxValue;
+    float currentValue;
+    float minValue;
+    float maxValue;
     int valueMaxCount;
     std::deque<float> values;
     bool visible;
-    Trace(int id, std::string tag, int minValue, int maxValue, int valueMaxCount);
+    Trace(int id, std::string tag, float minValue, float maxValue, int valueMaxCount);
     virtual ~Trace() = default; // make it polymorphic
   };
 
@@ -45,10 +45,10 @@ private:
 public:
   Metrics(std::pair<int, int> viewport);
   void render(Text &text);
-  int addGraph(std::string tag, int minVal, int maxVal, int valuesMaxCount);
+  int addGraph(std::string tag, float minVal, float maxVal, int valuesMaxCount);
   void setGraphVisibility(int graphId, bool visible);
-  void updateGraphValues(int graphId, std::vector<int> values);
-  void updateGraphValue(int graphId, int value);
+  void updateGraphValues(int graphId, std::vector<float> values);
+  void updateGraphValue(int graphId, float value);
 };
 
 #endif // _FRAMERATE_H_
